fix: Use View Engine LS for projects < v9
`ngcc` does not work correctly for Angular projects < version 9, so
automatically switch to View Engine LS when such projects are detected in
the workspace.
